Lots of long-standing bugfixes and small updates.

+++ zerorpc/heartbeat.py
@@ -31,11 +31,13 @@ import gevent.local
 import gevent.lock
 
 from .exceptions import *  # noqa
+from .channel_base import ChannelBase
 
 
-class HeartBeatOnChannel(object):
+class HeartBeatOnChannel(ChannelBase):
 
     def __init__(self, channel, freq=5, passive=False):
+        self._closed = False
         self._channel = channel
         self._heartbeat_freq = freq
         self._input_queue = gevent.queue.Channel()
@@ -49,13 +51,15 @@ class HeartBeatOnChannel(object):
             self._start_heartbeat()
 
     @property
-    def recv_is_available(self):
-        return self._channel.recv_is_available
+    def recv_is_supported(self):
+        return self._channel.recv_is_supported
 
-    def __del__(self):
-        self.close()
+    @property
+    def emit_is_supported(self):
+        return self._channel.emit_is_supported
 
     def close(self):
+        self._closed = True
         if self._heartbeat_task is not None:
             self._heartbeat_task.kill()
             self._heartbeat_task = None
@@ -73,13 +77,14 @@ class HeartBeatOnChannel(object):
                 self._remote_last_hb = time.time()
             if time.time() > self._remote_last_hb + self._heartbeat_freq * 2:
                 self._lost_remote = True
-                gevent.kill(self._parent_coroutine,
-                        self._lost_remote_exception())
+                if not self._closed:
+                    gevent.kill(self._parent_coroutine,
+                            self._lost_remote_exception())
                 break
             self._channel.emit('_zpc_hb', (0,))  # 0 -> compat with protocol v2
 
     def _start_heartbeat(self):
-        if self._heartbeat_task is None and self._heartbeat_freq is not None:
+        if self._heartbeat_task is None and self._heartbeat_freq is not None and not self._closed:
             self._heartbeat_task = gevent.spawn(self._heartbeat)
 
     def _recver(self):
@@ -100,30 +105,23 @@ class HeartBeatOnChannel(object):
         return LostRemote('Lost remote after {0}s heartbeat'.format(
             self._heartbeat_freq * 2))
 
-    def create_event(self, name, args, xheader=None):
+    def new_event(self, name, args, header=None):
         if self._compat_v2 and name == '_zpc_more':
             name = '_zpc_hb'
-        return self._channel.create_event(name, args, xheader)
+        return self._channel.new_event(name, args, header)
 
-    def emit_event(self, event):
+    def emit_event(self, event, timeout=None):
         if self._lost_remote:
             raise self._lost_remote_exception()
-        self._channel.emit_event(event)
-
-    def emit(self, name, args, xheader=None):
-        event = self.create_event(name, args, xheader)
-        self.emit_event(event)
+        self._channel.emit_event(event, timeout)
 
     def recv(self, timeout=None):
         if self._lost_remote:
             raise self._lost_remote_exception()
-
         try:
-            event = self._input_queue.get(timeout=timeout)
+            return self._input_queue.get(timeout=timeout)
         except gevent.queue.Empty:
             raise TimeoutExpired(timeout)
-
-        return event
 
     @property
     def channel(self):
